requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0
ollama==0.3.3
chromadb==0.5.20
pypdf==5.1.0
python-docx==1.1.2
langchain-text-splitters==0.3.0
watchdog==6.0.0
python-dotenv==1.0.1
httpx==0.27.2
typer==0.13.1
structlog==24.4.0
python-multipart==0.0.12
